2010-10-12  Chris Marrin  <cmarrin@apple.com>
        Reviewed by Darin Adler.

        Manage DrawingBuffer lifetime in GraphicsContext3D
        https://bugs.webkit.org/show_bug.cgi?id=47501

        GraphicsContext3D and DrawingBuffer are now Refcounted. DrawingBuffer
        has a RefPtr to GraphicsContext3D to ensure the proper lifetime.
        DrawingBuffer is now created by GraphicsContext3D. Also DrawingBuffer
        refers to GraphicsContext3D rather than SharedGraphicsContext3D.



Canonical link: https://commits.webkit.org/60213@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@69619 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ WebCore/platform/graphics/chromium/DrawingBufferChromium.cpp
@@ -48,7 +48,7 @@ struct DrawingBufferInternal {
 #endif
 };
 
-static unsigned generateColorTexture(SharedGraphicsContext3D* context, const IntSize& size)
+static unsigned generateColorTexture(GraphicsContext3D* context, const IntSize& size)
 {
     unsigned offscreenColorTexture = context->createTexture();
     if (!offscreenColorTexture)
@@ -66,13 +66,13 @@ static unsigned generateColorTexture(SharedGraphicsContext3D* context, const Int
 }
 
 
-DrawingBuffer::DrawingBuffer(SharedGraphicsContext3D* context, const IntSize& size, unsigned framebuffer)
+DrawingBuffer::DrawingBuffer(GraphicsContext3D* context, const IntSize& size)
     : m_context(context)
     , m_size(size)
-    , m_framebuffer(framebuffer)
+    , m_fbo(context->createFramebuffer())
     , m_internal(new DrawingBufferInternal)
 {
-    context->bindFramebuffer(framebuffer);
+    context->bindFramebuffer(GraphicsContext3D::FRAMEBUFFER, m_fbo);
     m_internal->offscreenColorTexture = generateColorTexture(context, size);
 }
 
@@ -82,14 +82,22 @@ DrawingBuffer::~DrawingBuffer()
     if (m_internal->platformLayer)
         m_internal->platformLayer->setDrawingBuffer(0);
 #endif
-    m_context->bindFramebuffer(m_framebuffer);
+
+    if (!m_context)
+        return;
+        
+    m_context->bindFramebuffer(GraphicsContext3D::FRAMEBUFFER, m_fbo);
     m_context->deleteTexture(m_internal->offscreenColorTexture);
-    m_context->deleteFramebuffer(m_framebuffer);
+
+    clear();
 }
 
 #if USE(ACCELERATED_COMPOSITING)
 void DrawingBuffer::publishToPlatformLayer()
 {
+    if (!m_context)
+        return;
+        
     if (m_callback)
         m_callback->willPublish();
     unsigned parentTexture = m_internal->platformLayer->textureId();
@@ -106,6 +114,9 @@ void DrawingBuffer::publishToPlatformLayer()
 
 void DrawingBuffer::reset(const IntSize& newSize)
 {
+    if (!m_context)
+        return;
+        
     if (m_size == newSize)
         return;
     m_size = newSize;
